+++ .example.env
@@ -1,7 +1,11 @@
+# database
 DB_HOST=localhost
 DB_USER=postgres
 DB_NAME=postgres
 DB_PASSWORD=1q2w3e4r
+DB_PORT=5432
+
+# kakao
 KAKAO_CLIENT_ID=
 KAKAO_REDIRECT_URL=
 KAKAO_REST_API_KEY=
