+++ src/common/guards/passport-auth.guard.ts
@@ -2,7 +2,7 @@ import { ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
 
-import { UserRole } from 'src/user/entities/user.type';
+import { UserRole } from 'src/entities';
 
 @Injectable()
 export class PassportAuthGuard extends AuthGuard('jwt') {
