+++ src/entities/group-map.entity.ts
@@ -1,17 +1,14 @@
 import {
   Collection,
   Entity,
-  EntityRepositoryType,
   OneToMany,
   PrimaryKey,
   Property,
 } from '@mikro-orm/core';
 
-import { UserMap } from 'src/user-map/entities/user-map.entity';
+import { UserMap } from 'src/entities';
 
-import { MapRepository } from '../map.repository';
-
-@Entity({ repository: () => MapRepository, tableName: 'map' })
+@Entity({ tableName: 'map' })
 export class GroupMap {
   @PrimaryKey()
   id: string;
@@ -22,5 +19,9 @@ export class GroupMap {
   @OneToMany({ entity: () => UserMap, mappedBy: (userMap) => userMap.map })
   userMap = new Collection<UserMap>(this);
 
-  [EntityRepositoryType]?: MapRepository;
+  @Property()
+  createdAt: Date = new Date();
+
+  @Property({ onUpdate: () => new Date() })
+  updatedAt: Date = new Date();
 }
