+++ src/entities/group-map.repository.ts
@@ -1,0 +1,5 @@
+import { ExtendedEntityRepository } from 'src/common/helper/extended-repository.helper';
+
+import { GroupMap } from './group-map.entity';
+
+export class GroupMapRepository extends ExtendedEntityRepository<GroupMap> {}
