+++ src/entities/index.ts
@@ -1,0 +1,14 @@
+import { GroupMap } from './group-map.entity';
+import { UserMap } from './user-map.entity';
+import { User } from './user.entity';
+
+export * from './group-map.entity';
+export * from './group-map.repository';
+
+export * from './user-map.entity';
+export * from './user-map.repository';
+
+export * from './user.entity';
+export * from './user.repository';
+
+export const entities = [User, GroupMap, UserMap];
