+++ src/entities/user-map.repository.ts
@@ -1,5 +1,5 @@
 import { ExtendedEntityRepository } from 'src/common/helper/extended-repository.helper';
 
-import { UserMap } from './entities/user-map.entity';
+import { UserMap } from './user-map.entity';
 
 export class UserMapRepository extends ExtendedEntityRepository<UserMap> {}
