+++ src/entities/user.repository.ts
@@ -1,5 +1,5 @@
 import { ExtendedEntityRepository } from 'src/common/helper/extended-repository.helper';
 
-import { User } from './entities/user.entity';
+import { User } from './user.entity';
 
 export class UserRepository extends ExtendedEntityRepository<User> {}
