+++ src/map/dtos/create-map.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 
 import { IsNotEmpty } from 'class-validator';
 
-import { GroupMap } from '../entities/map.entity';
+import { GroupMap } from 'src/entities';
 
 export class CreateMapDto implements Partial<GroupMap> {
   @ApiProperty()
