+++ src/map/dtos/update-map.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 
 import { IsOptional } from 'class-validator';
 
-import { GroupMap } from '../entities/map.entity';
+import { GroupMap } from 'src/entities';
 
 export class UpdateMapDto implements Partial<GroupMap> {
   @ApiProperty({ required: false })
