--- src/map/map.repository.ts
@@ -1,5 +1,0 @@
-import { ExtendedEntityRepository } from 'src/common/helper/extended-repository.helper';
-
-import { GroupMap } from './entities/map.entity';
-
-export class MapRepository extends ExtendedEntityRepository<GroupMap> {}
