+++ src/map/map.service.ts
@@ -2,17 +2,17 @@ import { Injectable } from '@nestjs/common';
 
 import { FilterQuery } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { wrap } from '@mikro-orm/postgresql';
+
+import { GroupMap, GroupMapRepository } from 'src/entities';
 
 import { CreateMapDto } from './dtos/create-map.dto';
 import { UpdateMapDto } from './dtos/update-map.dto';
-import { GroupMap } from './entities/map.entity';
-import { MapRepository } from './map.repository';
 
 @Injectable()
 export class MapService {
   constructor(
-    @InjectRepository(GroupMap) private readonly mapRepository: MapRepository,
+    @InjectRepository(GroupMap)
+    private readonly mapRepository: GroupMapRepository,
   ) {}
   async create(createMapDto: CreateMapDto) {
     const map: GroupMap = this.mapRepository.create(createMapDto);
@@ -29,7 +29,7 @@ export class MapService {
   }
   async update(id: string, updateMapDto: UpdateMapDto) {
     const map = await this.mapRepository.findOneOrFail(id);
-    wrap(map).assign(updateMapDto);
+    // TODO: 이거 다시 구현
     await this.mapRepository.persistAndFlush(map);
     return map;
   }
