+++ src/user/dtos/update-user.dto.ts
@@ -2,11 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 
 import { IsOptional } from 'class-validator';
 
-import { User } from '../entities/user.entity';
-import {
-  UserProviderValueType,
-  UserRoleValueType,
-} from '../entities/user.type';
+import { User, UserProviderValueType, UserRoleValueType } from 'src/entities';
 
 export class UpdateUserDto implements Partial<User> {
   @ApiProperty({ required: false })
