+++ src/user/user.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 
-import { User } from './entities/user.entity';
+import { User } from 'src/entities';
+
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
