+++ src/user/user.service.ts
@@ -2,12 +2,11 @@ import { Injectable } from '@nestjs/common';
 
 import { FilterQuery } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { wrap } from '@mikro-orm/postgresql';
+
+import { User, UserRepository } from 'src/entities';
 
 import { CreateUserDto } from './dtos/create-user.dto';
 import { UpdateUserDto } from './dtos/update-user.dto';
-import { User } from './entities/user.entity';
-import { UserRepository } from './user.repository';
 
 @Injectable()
 export class UserService {
@@ -21,7 +20,7 @@ export class UserService {
   }
 
   async findAll() {
-    const users = await this.userRepository.findAll();
+    const users = await this.userRepository.findAll({ populate: ['userMap'] });
     return users;
   }
 
@@ -32,7 +31,8 @@ export class UserService {
 
   async update(id: number, updateUserDto: UpdateUserDto) {
     const user = await this.userRepository.findOneOrFail(id);
-    wrap(user).assign(updateUserDto);
+    // TODO: 다시 구현
+    // (wrap(user) as any).assign(updateUserDto);
     await this.userRepository.persistAndFlush(user);
     return user;
   }
